move typedoc into commerce-sdk-react, and test run the script
