+++ package.json
@@ -9,15 +9,11 @@
     "commander": "^2.20.0",
     "lerna": "^3.20.2",
     "semver": "^7.3.2",
-    "shelljs": "^0.8.5",
-    "typedoc": "^0.23.5",
-    "typescript": "^4.7.2",
-    "typedoc-plugin-external-module-map": "1.3.2"
+    "shelljs": "^0.8.5"
   },
   "scripts": {
     "preinstall": "node ./scripts/check-version.js",
     "postinstall": "node ./scripts/bootstrap.js",
-    "build:docs": "typedoc",
     "lerna": "lerna",
     "pretest": "npm run lint",
     "test": "lerna run --stream --concurrency=1 test",
