+++ packages/commerce-sdk-react/typedoc.json
@@ -1,0 +1,6 @@
+{
+  "out": "docs",
+  "entryPoints": [
+    "src/index.ts"
+  ]
+}
