--- typedoc.json
@@ -1,8 +1,0 @@
-{
-  "out": "docs",
-  "entryPoints": [
-    "packages/commerce-sdk-react"
-  ],
-  "entryPointStrategy": "packages",
-  "logLevel": "Verbose"
-}
